Move common fns from converter out to virt-launcher/util.go
