+++ pkg/virt-launcher-libvirt-qemu/virtwrap/converter/vcpu/numa_placement_test.go
@@ -10,6 +10,7 @@ import (
 	v1 "kubevirt.io/api/core/v1"
 
 	cmdv1 "kubevirt.io/kubevirt/pkg/handler-launcher-com/cmd/v1"
+	virtlauncher "kubevirt.io/kubevirt/pkg/virt-launcher"
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/api"
 )
 
@@ -37,7 +38,7 @@ var _ = Describe("NumaPlacement", func() {
 				EmulatorPin: nil,
 			},
 		}
-		givenSpec.Memory, err = QuantityToByte(resource.MustParse("64Mi"))
+		givenSpec.Memory, err = virtlauncher.QuantityToByte(resource.MustParse("64Mi"))
 		Expect(err).ToNot(HaveOccurred())
 		givenTopology = &cmdv1.Topology{
 			NumaCells: []*cmdv1.Cell{
@@ -131,7 +132,7 @@ var _ = Describe("NumaPlacement", func() {
 		It("should detect if not enough memory is requested", func() {
 			var err error
 			memory := resource.MustParse("2Mi")
-			givenSpec.Memory, err = QuantityToByte(memory)
+			givenSpec.Memory, err = virtlauncher.QuantityToByte(memory)
 			Expect(err).ToNot(HaveOccurred())
 			givenVMI.Spec.Domain.Memory.Guest = &memory
 			Expect(numaMapping(givenVMI, givenSpec, givenTopology)).ToNot(Succeed())
@@ -139,7 +140,7 @@ var _ = Describe("NumaPlacement", func() {
 
 		It("should detect not divisable hugepages and shuffle the memory", func() {
 			var err error
-			givenSpec.Memory, err = QuantityToByte(resource.MustParse("66Mi"))
+			givenSpec.Memory, err = virtlauncher.QuantityToByte(resource.MustParse("66Mi"))
 			Expect(err).ToNot(HaveOccurred())
 			givenSpec.CPUTune.VCPUPin = append(givenSpec.CPUTune.VCPUPin, api.CPUTuneVCPUPin{
 				VCPU: 4, CPUSet: "40",
